# Update air_rework_defs.lua