T1 fighters +10% speed, +20% cost
T2 fighters +10% speed, +20% cost
Highwind 2.7s->6s reloadtime
Nighthawk 1.15s->2s reloadtime, 1050->850 range
Heavy fighters +10% speed.

Scout planes +10% speed
Finch, Blink +20% sight

T1 transports +7% hp

Bombers +10% speed, +20% cost

Gunships +13% speed, -13% hp (excluding Dragon, Shuriken).  Wasp, Roughneck +20% weapon damage
Hornet restored to vanilla stats, but with faster turning, more range (550), and more AoE (64) Banshee got 24 AoE, to help it hit tiny moving targets like ticks

Ferret / SAM -30% e/m cost
Flak turrets -10% e/m cost

<!--
PR Template! Please make sure to give your PR a relevant title so a squash merge remains descriptive
If any commented sections are not relevant to this PR, remove them.
Please fill out the uncommented sections with any relevant information.
-->

### Work done
<!--
Describe the changes or additions made in this PR, and why they
are necessary or important. If there is unusual complexity in the
code or functionality, please explain it so reviewers can understand.
-->

<!-- If relevant
#### Addresses Issue(s)
- Issue URL
-->

<!-- If relevant
#### Setup
Describe any setup requirements to test this work (Specific settings, widgets, etc))
-->

#### Test steps
- [ ] Write the steps needed to test out the changes. Include the expected result.

<!-- If relevant
### Screenshots:
If you're making visible changes, add before/after screenshots or videos of the major
changes so it's easier for reviewers to see what is different in this PR

#### BEFORE:
(screenshot from master)

#### AFTER:
(screenshot from branch)
-->